t	plit astconv's error report code in check functions to mod errors.
Move some error report codes to mod `astconv/errors.rs`
